requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.4
cocos2d==0.6.8
colorama==0.4.3
Django==3.1.4
image==1.5.33
isort==5.6.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pycparser==2.20
pyglet==1.5.8
pylint==2.6.0
pymunk==5.7.0
PyOpenGL==3.1.1a1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
wincertstore==0.2
wrapt==1.12.1
